fix(Updates): Add SSR by moving query to page
